global settings context, provider, hook
