+++ src/Contexts/GlobalSettingsContext.tsx
@@ -1,0 +1,10 @@
+import { createContext } from 'react';
+
+export type GlobalSettingsContextType = {
+  isFullscreen: boolean;
+  enableFullscreen: () => void;
+  disableFullscreen: () => void;
+};
+
+export const GlobalSettingsContext =
+  createContext<GlobalSettingsContextType | null>(null);
